@@ -78,16 +78,16 @@
         "deploy:ci": "ng deploy --no-build --message=\"Release $npm_package_name (v$npm_package_version) on gh-pages\""
     },
     "dependencies": {
-        "@angular/animations": "~13.3.11",
-        "@angular/common": "~13.3.11",
-        "@angular/compiler": "~13.3.11",
-        "@angular/core": "~13.3.11",
-        "@angular/forms": "~13.3.11",
-        "@angular/localize": "~13.3.11",
-        "@angular/platform-browser": "~13.3.11",
-        "@angular/platform-browser-dynamic": "~13.3.11",
-        "@angular/platform-server": "~13.3.11",
-        "@angular/router": "~13.3.11",
+        "@angular/animations": "^14.2.3",
+        "@angular/common": "^14.2.3",
+        "@angular/compiler": "^14.2.3",
+        "@angular/core": "^14.2.3",
+        "@angular/forms": "^14.2.3",
+        "@angular/localize": "^14.2.3",
+        "@angular/platform-browser": "^14.2.3",
+        "@angular/platform-browser-dynamic": "^14.2.3",
+        "@angular/platform-server": "^14.2.3",
+        "@angular/router": "^14.2.3",
         "@ctrl/ngx-codemirror": "^5.1.1",
         "@fortawesome/angular-fontawesome": "^0.10.2",
         "@fortawesome/fontawesome-svg-core": "^6.2.0",
@@ -113,14 +113,14 @@
         "zone.js": "~0.11.8"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~13.3.9",
+        "@angular-devkit/build-angular": "^14.2.3",
         "@angular-eslint/builder": "^13.5.0",
         "@angular-eslint/eslint-plugin": "^13.5.0",
         "@angular-eslint/eslint-plugin-template": "^13.5.0",
         "@angular-eslint/schematics": "^13.5.0",
         "@angular-eslint/template-parser": "^13.5.0",
-        "@angular/cli": "~13.3.9",
-        "@angular/compiler-cli": "~13.3.11",
+        "@angular/cli": "^14.2.3",
+        "@angular/compiler-cli": "^14.2.3",
         "@commitlint/cli": "^17.1.2",
         "@commitlint/config-angular": "^17.1.0",
         "@compodoc/compodoc": "^1.1.19",
